Add support for pressing remote control buttons
- Move TV class out of __init__
- Support pressing defined remote control buttons
